Only run TestGetImageConfigStopSignal on Linux
... because the implementation requires Linux-only pkg/signal
